requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements-dev.in
#
attrs==21.4.0
    # via pytest
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via pip-tools
idna==3.3
    # via requests
iniconfig==1.1.1
    # via pytest
packaging==21.3
    # via pytest
pep517==0.12.0
    # via pip-tools
pip-tools==6.6.0
    # via -r requirements-dev.in
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pyparsing==3.0.8
    # via packaging
pytest==7.1.2
    # via
    #   -r requirements-dev.in
    #   pytest-django
pytest-django==4.5.2
    # via -r requirements-dev.in
requests==2.27.1
    # via requests-mock
requests-mock==1.9.3
    # via -r requirements-dev.in
six==1.16.0
    # via requests-mock
tomli==2.0.1
    # via
    #   pep517
    #   pytest
urllib3==1.26.9
    # via requests
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
